@@ -5,7 +5,7 @@ author: Onyemaechi Okafor <peerwaya@yahoo.com>
 homepage: https://github.com/peerwaya/flutter_call_kit
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
